+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v0.10.2 UNRELEASED
 
+### Added
+
+- DeepVar network (#923)
+
 ### Fixed
 
 - Fix learning scale schedule (#912)